backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
passlib==1.7.4
pip @ file:///croot/pip_1746204010231/work
pyasn1==0.4.8
pydantic==2.11.4
pydantic-core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uv==0.7.3
uvicorn==0.34.2
uvloop==0.21.0
